Extract eSIM functionality
