+++ android/src/main/java/com/reactnativesimcardsmanager/SimCardsManagerModule.java
@@ -2,16 +2,10 @@ package com.reactnativesimcardsmanager;
 
 import android.os.Build;
 import com.facebook.react.bridge.*;
-import android.app.PendingIntent;
-import android.content.Intent;
-import android.telephony.euicc.DownloadableSubscription;
-import android.content.IntentFilter;
 import android.content.Context;
-import android.content.BroadcastReceiver;
 import android.telephony.TelephonyManager;
 import android.telephony.SubscriptionInfo;
 import android.telephony.SubscriptionManager;
-import android.telephony.euicc.EuiccManager;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
@@ -27,22 +21,13 @@ import java.util.List;
 @ReactModule(name = SimCardsManagerModule.NAME)
 public class SimCardsManagerModule extends ReactContextBaseJavaModule {
   public static final String NAME = "SimCardsManager";
-  private final String ACTION_DOWNLOAD_SUBSCRIPTION = "download_subscription";
   private final ReactContext mReactContext;
-
-  @RequiresApi(api = Build.VERSION_CODES.P)
-  private EuiccManagerWrapper mgrWrapper;
+  private final EsimManager eSimMgr;
 
   public SimCardsManagerModule(ReactApplicationContext reactContext) {
     super(reactContext);
     mReactContext = reactContext;
-  }
-
-  @RequiresApi(api = Build.VERSION_CODES.P)
-  private void initMgr() {
-    if (mgrWrapper == null) {
-      mgrWrapper = new EuiccManagerWrapper(mReactContext);
-    }
+    eSimMgr = new EsimManager(reactContext);
   }
 
   @Override
@@ -104,31 +89,12 @@ public class SimCardsManagerModule extends ReactContextBaseJavaModule {
     promise.resolve(simCardsList);
   }
 
-  @RequiresApi(api = Build.VERSION_CODES.P)
   @ReactMethod
   public void isEsimSupported(Promise promise) {
-    initMgr();
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P && mgrWrapper.getMgr() != null) {
-      promise.resolve(mgrWrapper.getMgr().isEnabled());
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+      promise.resolve(eSimMgr.isEnabled());
     } else {
       promise.resolve(false);
-    }
-  }
-
-  @RequiresApi(api = Build.VERSION_CODES.P)
-  private void handleResolvableError(Promise promise, Intent intent) {
-    try {
-      // Resolvable error, attempt to resolve it by a user action
-      // FIXME: review logic of resolve functions
-      int resolutionRequestCode = 3;
-      PendingIntent callbackIntent = PendingIntent.getBroadcast(mReactContext, resolutionRequestCode,
-          new Intent(ACTION_DOWNLOAD_SUBSCRIPTION), PendingIntent.FLAG_UPDATE_CURRENT |
-              PendingIntent.FLAG_MUTABLE);
-
-      mgrWrapper.getMgr().startResolutionActivity(mReactContext.getCurrentActivity(), resolutionRequestCode, intent, callbackIntent);
-    } catch (Exception e) {
-      promise.reject("3", "EMBEDDED_SUBSCRIPTION_RESULT_RESOLVABLE_ERROR - Can't setup eSim du to Activity error "
-          + e.getLocalizedMessage());
     }
   }
 
@@ -141,68 +107,12 @@ public class SimCardsManagerModule extends ReactContextBaseJavaModule {
     }
   }
 
-  @RequiresApi(api = Build.VERSION_CODES.P)
   @ReactMethod
   public void setupEsim(ReadableMap config, Promise promise) {
-    initMgr();
-    if (android.os.Build.VERSION.SDK_INT < Build.VERSION_CODES.P) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+      eSimMgr.setupEsim(config, promise);
+    } else {
       promise.reject("0", "EuiccManager is not available or before Android 9 (API 28)");
     }
-
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P && mgrWrapper.getMgr() != null && !mgrWrapper.getMgr().isEnabled()) {
-      promise.reject("1", "The device doesn't support a cellular plan (EuiccManager is not available)");
-      return;
-    }
-
-//    if (!checkCarrierPrivileges()) {
-//      promise.reject("1", "No carrier privileges detected");
-//      return;
-//    }
-
-    BroadcastReceiver receiver = new BroadcastReceiver() {
-
-      @Override
-      public void onReceive(Context context, Intent intent) {
-        if (!ACTION_DOWNLOAD_SUBSCRIPTION.equals(intent.getAction())) {
-          promise.reject("3",
-              "Can't setup eSim du to wrong Intent:" + intent.getAction() + " instead of "
-                  + ACTION_DOWNLOAD_SUBSCRIPTION);
-          return;
-        }
-        int resultCode = getResultCode();
-        if (resultCode == EuiccManager.EMBEDDED_SUBSCRIPTION_RESULT_RESOLVABLE_ERROR && mgrWrapper.getMgr() != null) {
-          handleResolvableError(promise, intent);
-        } else if (resultCode == EuiccManager.EMBEDDED_SUBSCRIPTION_RESULT_OK) {
-          promise.resolve(true);
-        } else if (resultCode == EuiccManager.EMBEDDED_SUBSCRIPTION_RESULT_ERROR) {
-          // Embedded Subscription Error
-          promise.reject("2",
-              "EMBEDDED_SUBSCRIPTION_RESULT_ERROR - Can't add an Esim subscription");
-        } else {
-          // Unknown Error
-          promise.reject("3",
-              "Can't add an Esim subscription due to unknown error, resultCode is:" + String.valueOf(resultCode));
-        }
-      }
-    };
-
-    mReactContext.registerReceiver(
-        receiver,
-        new IntentFilter(ACTION_DOWNLOAD_SUBSCRIPTION),
-        null,
-        null);
-
-    DownloadableSubscription sub = DownloadableSubscription.forActivationCode(
-        /* Passed from react side */
-        config.getString("confirmationCode"));
-
-    PendingIntent callbackIntent = PendingIntent.getBroadcast(
-        mReactContext,
-        0,
-        new Intent(ACTION_DOWNLOAD_SUBSCRIPTION),
-        PendingIntent.FLAG_UPDATE_CURRENT |
-            PendingIntent.FLAG_MUTABLE);
-
-    mgrWrapper.getMgr().downloadSubscription(sub, true, callbackIntent);
   }
 }
